+++ atmega_src/main.h
@@ -8,6 +8,7 @@
 
 #ifndef MAIN_H_
 #define MAIN_H_
+#include <avr/io.h>
 
 //void run_tests();
